backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
python-dotenv==1.1.0
supabase==2.15.2
groq==0.25.0
gunicorn==23.0.0
